solution + tests
Changed the tests file a lot to make it compatible with Set instead of Map. Nowhere in the instructions did it say to use a Map, so I went with what I was familiar with and that ended up putting me at a disadvantage. I'll look at the tests closer before I make the code next time.

At least I solved it thought :sweat_smile: